+++ src/AbstractClient.php
@@ -4,9 +4,16 @@ namespace EasyHTTP\Contracts;
 
 use EasyHTTP\Contracts\Contracts\HTTPClientContract;
 use EasyHTTP\Contracts\Contracts\HTTPClientAdapter;
+use EasyHTTP\Contracts\Contracts\HTTPClientEvent;
 use EasyHTTP\Contracts\Contracts\HTTPClientRequest;
 use EasyHTTP\Contracts\Contracts\HTTPClientResponse;
 use EasyHTTP\Contracts\Contracts\HTTPStreamResponse;
+use EasyHTTP\Contracts\Events\RequestFailed;
+use EasyHTTP\Contracts\Events\RequestStarted;
+use EasyHTTP\Contracts\Events\RequestSucceeded;
+use EasyHTTP\Contracts\Events\StreamSucceeded;
+use Psr\EventDispatcher\EventDispatcherInterface;
+use Throwable;
 
 abstract class AbstractClient implements HTTPClientContract
 {
@@ -14,6 +21,7 @@ abstract class AbstractClient implements HTTPClientContract
     protected HTTPClientRequest $request;
     protected HTTPClientFactory $factory;
     protected $handler;
+    protected ?EventDispatcherInterface $eventDispatcher = null;
 
     public function __construct(HTTPClientFactory $factory)
     {
@@ -28,7 +36,18 @@ abstract class AbstractClient implements HTTPClientContract
     public function call(string $method, string $uri): HTTPClientResponse
     {
         $request = $this->factory->createRequest($method, $uri);
-        return $this->getAdapter()->request($request);
+
+        $this->emit(new RequestStarted($request, ['operation' => 'call']));
+
+        try {
+            $response = $this->getAdapter()->request($request);
+            $this->emit(new RequestSucceeded($response, ['operation' => 'call']));
+
+            return $response;
+        } catch (Throwable $exception) {
+            $this->emit(new RequestFailed($request, $exception, ['operation' => 'call']));
+            throw $exception;
+        }
     }
 
     public function prepareRequest(string $method, string $uri): HTTPClientRequest
@@ -46,14 +65,55 @@ abstract class AbstractClient implements HTTPClientContract
         return $this;
     }
 
+    public function withEventDispatcher(EventDispatcherInterface $eventDispatcher): self
+    {
+        $this->eventDispatcher = $eventDispatcher;
+
+        return $this;
+    }
+
     public function execute(): HTTPClientResponse
     {
-        return $this->getAdapter()->request($this->request);
+        $this->emit(new RequestStarted($this->request, ['operation' => 'execute']));
+
+        try {
+            $response = $this->getAdapter()->request($this->request);
+            $this->emit(new RequestSucceeded($response, ['operation' => 'execute']));
+
+            return $response;
+        } catch (Throwable $exception) {
+            $this->emit(new RequestFailed($this->request, $exception, ['operation' => 'execute']));
+            throw $exception;
+        }
     }
 
     public function stream(): HTTPStreamResponse
     {
-        return $this->getAdapter()->stream($this->request);
+        $this->emit(new RequestStarted($this->request, ['operation' => 'stream']));
+
+        try {
+            $response = $this->getAdapter()->stream($this->request);
+            $this->emit(new StreamSucceeded($response, ['operation' => 'stream']));
+
+            return $response;
+        } catch (Throwable $exception) {
+            $this->emit(new RequestFailed($this->request, $exception, ['operation' => 'stream']));
+            throw $exception;
+        }
+    }
+
+    protected function emit(HTTPClientEvent $event): void
+    {
+        if (!$this->hasEventDispatcher()) {
+            return;
+        }
+
+        $this->eventDispatcher->dispatch($event);
+    }
+
+    protected function hasEventDispatcher(): bool
+    {
+        return $this->eventDispatcher !== null;
     }
 
     protected function getAdapter(): HTTPClientAdapter