+++ src/Contracts/HTTPClientEvent.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EasyHTTP\Contracts\Contracts;
+
+interface HTTPClientEvent
+{
+    public function getName(): string;
+    public function getContext(): array;
+}